Updated widescreen text
